+++ pom.xml
@@ -66,6 +66,16 @@
             <version>2.8.6</version>
         </dependency>
         <dependency>
+            <groupId>com.github.michaelruocco</groupId>
+            <artifactId>jackson-json-api</artifactId>
+            <version>2.0.2</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jsr310</artifactId>
+            <version>2.11.0</version>
+        </dependency>
+        <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>mongo-java-driver</artifactId>
             <version>2.12.3</version>
