--- src/main/java/de/ccd/groupprio/integration/api/dto/PrioDto.java
@@ -1,7 +1,0 @@
-package de.ccd.groupprio.integration.api.dto;
-
-import java.util.List;
-
-public class PrioDto {
-    public List<String> items;
-}
