+++ src/main/java/de/ccd/groupprio/integration/project/create/CreatedProjectResponse.java
@@ -1,0 +1,8 @@
+package de.ccd.groupprio.integration.project.create;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+class CreatedProjectResponse {
+    private final String id;
+}
