+++ src/main/java/de/ccd/groupprio/integration/project/get/GetAllProjectsQuery.java
@@ -1,0 +1,13 @@
+package de.ccd.groupprio.integration.project.get;
+
+import lombok.RequiredArgsConstructor;
+import spark.Request;
+
+@RequiredArgsConstructor
+class GetAllProjectsQuery {
+    final String clientId;
+
+    static GetAllProjectsQuery from(Request req){
+        return new GetAllProjectsQuery(req.headers("clientId"));
+    }
+}
