+++ src/main/java/de/ccd/groupprio/integration/project/get/GetOneProjectProcessor.java
@@ -1,0 +1,17 @@
+package de.ccd.groupprio.integration.project.get;
+
+import de.ccd.groupprio.repository.project.ProjectRepository;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+class GetOneProjectProcessor {
+
+    final ProjectRepository projectRepository;
+
+    ProjectResponse process(GetOneProjectQuery qry) {
+        var project = projectRepository
+                .getByProjectId(qry.projectId);
+        return ProjectResponse.map(project);
+    }
+
+}
