+++ src/main/java/de/ccd/groupprio/integration/project/state/ProjectStateQuery.java
@@ -1,0 +1,13 @@
+package de.ccd.groupprio.integration.project.state;
+
+import lombok.RequiredArgsConstructor;
+import spark.Request;
+
+@RequiredArgsConstructor
+class ProjectStateQuery {
+    final String projectId;
+
+    static ProjectStateQuery from(Request req) {
+        return new ProjectStateQuery(req.params(":id"));
+    }
+}
