+++ src/main/java/de/ccd/groupprio/integration/project/state/ProjectStateResponse.java
@@ -1,0 +1,12 @@
+package de.ccd.groupprio.integration.project.state;
+
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+class ProjectStateResponse {
+    final String title;
+    final List<String> weightedItems;
+    final int submissionCount;
+}
