+++ src/main/java/de/ccd/groupprio/integration/submit/SubmitResponse.java
@@ -1,0 +1,8 @@
+package de.ccd.groupprio.integration.submit;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+class SubmitResponse {
+    final boolean success;
+}
