--- src/main/java/de/ccd/groupprio/repository/SubmissionRepository.java
@@ -1,16 +1,0 @@
-package de.ccd.groupprio.repository;
-
-import de.ccd.groupprio.domain.data.Submission;
-
-import java.util.List;
-
-public interface SubmissionRepository {
-
-    List<Submission> findForProjectId(String projectId);
-
-    void save(String projectId, String clientId, Submission submission);
-
-    int getSubmissionCount(String projectId);
-
-    boolean hasClientSubmitted(String projectId, String clientId);
-}
