+++ src/main/java/de/ccd/groupprio/repository/weight/WeightRepository.java
@@ -1,0 +1,12 @@
+package de.ccd.groupprio.repository.weight;
+
+import de.ccd.groupprio.domain.data.WeightedItem;
+
+import java.util.List;
+
+public interface WeightRepository {
+
+    List<WeightedItem> findForProjectId(String projectId);
+
+    void save(String projectId, List<WeightedItem> items);
+}
